# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkflowRunCreateRequest": {
        "additionalProperties": false,
        "description": "Request a manifest-derived workflow plan for an active session.",
        "properties": {
          "workflow_id": {
            "maxLength": 16,
            "minLength": 1,
            "pattern": "^W-[A-Z]$",
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id"
        ],
        "title": "WorkflowRunCreateRequest",
        "type": "object"
      },
      "WorkflowRunListResponse": {
        "additionalProperties": false,
        "description": "List response for tenant-scoped workflow runs.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkflowRunResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "limit",
          "offset"
        ],
        "title": "WorkflowRunListResponse",
        "type": "object"
      },
      "WorkflowRunResponse": {
        "additionalProperties": false,
        "description": "Safe workflow run projection derived only from the manifest.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowRunStatus"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStepRunResponse"
            },
            "title": "Steps",
            "type": "array"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          },
          "workflow_slug": {
            "title": "Workflow Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "session_id",
          "workflow_id",
          "workflow_slug",
          "status",
          "created_at",
          "updated_at",
          "steps"
        ],
        "title": "WorkflowRunResponse",
        "type": "object"
      },
      "WorkflowRunStatus": {
        "description": "Lifecycle for a workflow run before execution gates open.",
        "enum": [
          "planned",
          "cancelled"
        ],
        "title": "WorkflowRunStatus",
        "type": "string"
      },
      "WorkflowStepRunResponse": {
        "additionalProperties": false,
        "description": "Safe workflow step projection without engine execution state.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "module_id": {
            "title": "Module Id",
            "type": "string"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowStepRunStatus"
          },
          "step_index": {
            "title": "Step Index",
            "type": "integer"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workflow_run_id": {
            "format": "uuid",
            "title": "Workflow Run Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "workflow_run_id",
          "step_index",
          "module_id",
          "phase",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "WorkflowStepRunResponse",
        "type": "object"
      },
      "WorkflowStepRunStatus": {
        "description": "Lifecycle for manifest-derived workflow step runs before execution gates open.",
        "enum": [
          "planned",
          "blocked"
        ],
        "title": "WorkflowStepRunStatus",
        "type": "string"
      }
    }
  },
  "info": {
    "description": "Controlled ground-zero rebuild API skeleton.",
    "title": "radiquant-v5 API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/sessions/{session_id}/workflow-runs": {
      "get": {
        "description": "List manifest-derived workflow runs for a tenant-owned session.",
        "operationId": "listSessionWorkflowRuns",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Session Workflow Runs",
        "tags": [
          "workflows"
        ]
      },
      "post": {
        "description": "Create a manifest-derived workflow plan for an active tenant-owned session.",
        "operationId": "createWorkflowRun",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowRunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workflow Run",
        "tags": [
          "workflows"
        ]
      }
    }
  }
}